requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-googlesearch==0.2
idna==3.2
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.1
telebot==0.0.4
urllib3==1.26.6
